backend/requirements.txt
fastapi
uvicorn
python-multipart  # For form data if needed (good practice)
google-genai
pydantic